create utils folder and move AuthenticationButton  to utils
